2024/05/19
Fixed comment lines, some bugs and README.md

+++ Exponential_Logarithmic/ft_logb.c
@@ -3,21 +3,25 @@
 /*                                                        :::      ::::::::   */
 /*   ft_logb.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hdeniz <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: hdeniz <Discord:@teomandeniz>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/01 20:45:44 by hdeniz            #+#    #+#             */
-/*   Updated: 2023/03/01 20:45:45 by hdeniz           ###   ########.fr       */
+/*   Updated: 2024/05/18 ??:??:?? by hdeniz           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-/* INCLUDES */
-#include "../ft_math.h"
-/* INCLUDES */
+/* **************************** [v] INCLUDES [v] **************************** */
+#include "../ft_math.h" /*
+#    int ft_isnan(double);
+#    int ft_isinf(double);
+# double ft_frexp(double, long *);
+#        */
+/* **************************** [^] INCLUDES [^] **************************** */
 
 double
 	ft_logb(register double x)
 {
-	int	exponent;
+	long	exponent;
 
 	if (x == 0.0)
 		return ((-1.0 / 0.0));
@@ -26,5 +30,5 @@ double
 	if (ft_isinf(x))
 		return ((1.0 / 0.0));
 	ft_frexp(x, &exponent);
-	return ((double)exponent - 1);
+	return ((double)exponent - 1.0);
 }
