+++ Exponential_Logarithmic/ft_scalbn.c
@@ -3,19 +3,21 @@
 /*                                                        :::      ::::::::   */
 /*   ft_scalbn.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hdeniz <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: hdeniz <Discord:@teomandeniz>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/02 13:30:51 by hdeniz            #+#    #+#             */
-/*   Updated: 2023/03/02 13:30:53 by hdeniz           ###   ########.fr       */
+/*   Updated: 2024/05/18 ??:??:?? by hdeniz           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-/* INCLUDES */
-#include "../ft_math.h"
-/* INCLUDES */
+/* **************************** [v] INCLUDES [v] **************************** */
+#include "../ft_math.h" /*
+# double ft_ldexp(double, long);
+#        */
+/* **************************** [^] INCLUDES [^] **************************** */
 
 double
-	ft_scalbn(register double x, register int exponent)
+	ft_scalbn(register double x, register long exponent)
 {
 	return (ft_ldexp(x, exponent));
 }
